+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/send/media/SendMediaGroup.kt
@@ -8,12 +8,16 @@ import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.InputMedia.*
 import dev.inmo.tgbotapi.types.message.abstracts.MediaGroupMessage
 import dev.inmo.tgbotapi.types.message.abstracts.TelegramBotAPIMessageDeserializeOnlySerializerClass
+import dev.inmo.tgbotapi.utils.*
 import dev.inmo.tgbotapi.utils.throwRangeError
-import dev.inmo.tgbotapi.utils.toJsonWithoutNulls
 import kotlinx.serialization.*
 import kotlinx.serialization.builtins.ListSerializer
 import kotlinx.serialization.json.buildJsonArray
 
+const val rawSendingMediaGroupsWarning = "Media groups contains restrictions related to combinations of media" +
+    " types. Currently it is possible to combine photo + video OR audio OR documents"
+
+@RiskFeature(rawSendingMediaGroupsWarning)
 fun SendMediaGroup(
     chatId: ChatIdentifier,
     media: List<MediaGroupMemberInputMedia>,
@@ -51,6 +55,46 @@ fun SendMediaGroup(
         )
     }
 }
+
+/**
+ * Use this method to be sure that you are correctly sending playlist with audios
+ *
+ * @see InputMediaAudio
+ */
+@Suppress("NOTHING_TO_INLINE")
+inline fun SendPlaylist(
+    chatId: ChatIdentifier,
+    media: List<AudioMediaGroupMemberInputMedia>,
+    disableNotification: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null
+) = SendMediaGroup(chatId, media, disableNotification, replyToMessageId)
+
+/**
+ * Use this method to be sure that you are correctly sending documents media group
+ *
+ * @see InputMediaDocument
+ */
+@Suppress("NOTHING_TO_INLINE")
+inline fun SendDocumentsGroup(
+    chatId: ChatIdentifier,
+    media: List<DocumentMediaGroupMemberInputMedia>,
+    disableNotification: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null
+) = SendMediaGroup(chatId, media, disableNotification, replyToMessageId)
+
+/**
+ * Use this method to be sure that you are correctly sending visual media group
+ *
+ * @see InputMediaPhoto
+ * @see InputMediaVideo
+ */
+@Suppress("NOTHING_TO_INLINE")
+inline fun SendVisualMediaGroup(
+    chatId: ChatIdentifier,
+    media: List<VisualMediaGroupMemberInputMedia>,
+    disableNotification: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null
+) = SendMediaGroup(chatId, media, disableNotification, replyToMessageId)
 
 private val messagesListSerializer: KSerializer<List<MediaGroupMessage>>
     = ListSerializer(TelegramBotAPIMessageDeserializeOnlySerializerClass())