+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/InputMedia/MediaGroupMemberInputMedia.kt
@@ -18,3 +18,9 @@ internal fun <T> T.buildArguments(withSerializer: SerializationStrategy<T>) = ar
 interface MediaGroupMemberInputMedia : InputMedia, CaptionedOutput {
     fun serialize(format: StringFormat): String
 }
+
+interface AudioMediaGroupMemberInputMedia: MediaGroupMemberInputMedia
+interface DocumentMediaGroupMemberInputMedia: MediaGroupMemberInputMedia
+
+@Serializable(MediaGroupMemberInputMediaSerializer::class)
+interface VisualMediaGroupMemberInputMedia : MediaGroupMemberInputMedia