+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/abstracts/MediaGroupContent.kt
@@ -6,3 +6,7 @@ import dev.inmo.tgbotapi.types.InputMedia.MediaGroupMemberInputMedia
 interface MediaGroupContent : MediaContent, CaptionedInput {
     fun toMediaGroupMemberInputMedia(): MediaGroupMemberInputMedia
 }
+
+interface VisualMediaGroupContent : MediaGroupContent
+interface AudioMediaGroupContent : MediaGroupContent
+interface DocumentMediaGroupContent : MediaGroupContent