+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/AudioContent.kt
@@ -1,18 +1,17 @@
 package dev.inmo.tgbotapi.types.message.content.media
 
-import dev.inmo.tgbotapi.CommonAbstracts.CaptionedInput
 import dev.inmo.tgbotapi.CommonAbstracts.TextPart
 import dev.inmo.tgbotapi.requests.abstracts.Request
 import dev.inmo.tgbotapi.requests.send.media.SendAudio
 import dev.inmo.tgbotapi.types.ChatIdentifier
-import dev.inmo.tgbotapi.types.InputMedia.InputMediaAudio
+import dev.inmo.tgbotapi.types.InputMedia.*
 import dev.inmo.tgbotapi.types.MessageIdentifier
 import dev.inmo.tgbotapi.types.ParseMode.HTMLParseMode
 import dev.inmo.tgbotapi.types.ParseMode.MarkdownV2
 import dev.inmo.tgbotapi.types.buttons.KeyboardMarkup
 import dev.inmo.tgbotapi.types.files.AudioFile
 import dev.inmo.tgbotapi.types.message.abstracts.ContentMessage
-import dev.inmo.tgbotapi.types.message.content.abstracts.MediaContent
+import dev.inmo.tgbotapi.types.message.content.abstracts.*
 import dev.inmo.tgbotapi.utils.toHtmlCaptions
 import dev.inmo.tgbotapi.utils.toMarkdownV2Captions
 
@@ -20,7 +19,7 @@ data class AudioContent(
     override val media: AudioFile,
     override val caption: String? = null,
     override val captionEntities: List<TextPart> = emptyList()
-) : MediaContent, CaptionedInput {
+) : AudioMediaGroupContent {
     override fun createResend(
         chatId: ChatIdentifier,
         disableNotification: Boolean,
@@ -40,13 +39,10 @@ data class AudioContent(
         replyMarkup
     )
 
-    override fun asInputMedia(): InputMediaAudio = InputMediaAudio(
-        media.fileId,
-        toMarkdownV2Captions().firstOrNull(),
-        MarkdownV2,
-        media.duration,
-        media.performer,
-        media.title,
-        media.thumb ?.fileId
+    override fun toMediaGroupMemberInputMedia(): InputMediaAudio = asInputMedia()
+
+    override fun asInputMedia(): InputMediaAudio = media.toInputMediaAudio(
+        toHtmlCaptions().firstOrNull(),
+        HTMLParseMode
     )
 }