+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/DocumentContent.kt
@@ -5,14 +5,15 @@ import dev.inmo.tgbotapi.CommonAbstracts.TextPart
 import dev.inmo.tgbotapi.requests.abstracts.Request
 import dev.inmo.tgbotapi.requests.send.media.SendDocument
 import dev.inmo.tgbotapi.types.ChatIdentifier
-import dev.inmo.tgbotapi.types.InputMedia.InputMediaDocument
+import dev.inmo.tgbotapi.types.InputMedia.*
 import dev.inmo.tgbotapi.types.MessageIdentifier
 import dev.inmo.tgbotapi.types.ParseMode.HTMLParseMode
 import dev.inmo.tgbotapi.types.ParseMode.MarkdownV2
 import dev.inmo.tgbotapi.types.buttons.KeyboardMarkup
 import dev.inmo.tgbotapi.types.files.DocumentFile
+import dev.inmo.tgbotapi.types.files.asDocumentFile
 import dev.inmo.tgbotapi.types.message.abstracts.ContentMessage
-import dev.inmo.tgbotapi.types.message.content.abstracts.MediaContent
+import dev.inmo.tgbotapi.types.message.content.abstracts.*
 import dev.inmo.tgbotapi.utils.toHtmlCaptions
 import dev.inmo.tgbotapi.utils.toMarkdownV2Captions
 
@@ -20,7 +21,7 @@ data class DocumentContent(
     override val media: DocumentFile,
     override val caption: String? = null,
     override val captionEntities: List<TextPart> = emptyList()
-) : MediaContent, CaptionedInput {
+) : DocumentMediaGroupContent {
     override fun createResend(
         chatId: ChatIdentifier,
         disableNotification: Boolean,
@@ -37,10 +38,22 @@ data class DocumentContent(
         replyMarkup
     )
 
-    override fun asInputMedia(): InputMediaDocument = InputMediaDocument(
-        media.fileId,
-        toMarkdownV2Captions().firstOrNull(),
-        MarkdownV2,
-        media.thumb ?.fileId
+    override fun toMediaGroupMemberInputMedia(): InputMediaDocument = asInputMedia()
+
+    override fun asInputMedia(): InputMediaDocument = media.toInputMediaDocument(
+        toHtmlCaptions().firstOrNull(),
+        HTMLParseMode
+    )
+}
+
+@Suppress("NOTHING_TO_INLINE")
+inline fun MediaContent.asDocumentContent() = when (this) {
+    is CaptionedInput -> DocumentContent(
+        media.asDocumentFile(),
+        caption,
+        captionEntities
+    )
+    else -> DocumentContent(
+        media.asDocumentFile()
     )
 }