+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/PhotoContent.kt
@@ -4,16 +4,14 @@ import dev.inmo.tgbotapi.CommonAbstracts.TextPart
 import dev.inmo.tgbotapi.requests.abstracts.Request
 import dev.inmo.tgbotapi.requests.send.media.SendPhoto
 import dev.inmo.tgbotapi.types.ChatIdentifier
-import dev.inmo.tgbotapi.types.InputMedia.InputMediaPhoto
-import dev.inmo.tgbotapi.types.InputMedia.MediaGroupMemberInputMedia
+import dev.inmo.tgbotapi.types.InputMedia.*
 import dev.inmo.tgbotapi.types.MessageIdentifier
 import dev.inmo.tgbotapi.types.ParseMode.HTMLParseMode
 import dev.inmo.tgbotapi.types.ParseMode.MarkdownV2
 import dev.inmo.tgbotapi.types.buttons.KeyboardMarkup
 import dev.inmo.tgbotapi.types.files.*
 import dev.inmo.tgbotapi.types.message.abstracts.ContentMessage
-import dev.inmo.tgbotapi.types.message.content.abstracts.MediaCollectionContent
-import dev.inmo.tgbotapi.types.message.content.abstracts.MediaGroupContent
+import dev.inmo.tgbotapi.types.message.content.abstracts.*
 import dev.inmo.tgbotapi.utils.toHtmlCaptions
 import dev.inmo.tgbotapi.utils.toMarkdownV2Captions
 
@@ -21,7 +19,7 @@ data class PhotoContent(
     override val mediaCollection: Photo,
     override val caption: String? = null,
     override val captionEntities: List<TextPart> = emptyList()
-) : MediaCollectionContent<PhotoSize>, MediaGroupContent {
+) : MediaCollectionContent<PhotoSize>, VisualMediaGroupContent {
     override val media: PhotoSize = mediaCollection.biggest() ?: throw IllegalStateException("Can't locate any photo size for this content")
 
     override fun createResend(
@@ -39,15 +37,10 @@ data class PhotoContent(
         replyMarkup
     )
 
-    override fun toMediaGroupMemberInputMedia(): MediaGroupMemberInputMedia = InputMediaPhoto(
-        media.fileId,
+    override fun toMediaGroupMemberInputMedia(): InputMediaPhoto = asInputMedia()
+
+    override fun asInputMedia(): InputMediaPhoto = media.toInputMediaPhoto(
         toHtmlCaptions().firstOrNull(),
         HTMLParseMode
-    )
-
-    override fun asInputMedia(): InputMediaPhoto = InputMediaPhoto(
-        media.fileId,
-        toMarkdownV2Captions().firstOrNull(),
-        MarkdownV2
     )
 }